Move loading-instructions into src/preact
