+++ src/preact/loading-instructions.js
